+++ src/main/java/kitchenpos/orders/delivery_orders/domain/KitchenridersClient.java
@@ -1,0 +1,8 @@
+package kitchenpos.orders.delivery_orders.domain;
+
+import java.math.BigDecimal;
+import java.util.UUID;
+
+public interface KitchenridersClient {
+    void requestDelivery(UUID orderId, BigDecimal amount, String deliveryAddress);
+}
