+++ src/main/java/kitchenpos/orders/domain/OrderRepository.java
@@ -1,0 +1,18 @@
+package kitchenpos.orders.domain;
+
+import kitchenpos.orders.eatin_orders.domain.OrderTable;
+
+import java.util.List;
+import java.util.Optional;
+import java.util.UUID;
+
+public interface OrderRepository {
+    Order save(Order order);
+
+    Optional<Order> findById(UUID id);
+
+    List<Order> findAll();
+
+    boolean existsByOrderTableAndStatusNot(OrderTable orderTable, OrderStatus status);
+}
+
