+++ src/main/java/kitchenpos/orders/domain/OrderStatus.java
@@ -1,0 +1,5 @@
+package kitchenpos.orders.domain;
+
+public enum OrderStatus {
+    WAITING, ACCEPTED, SERVED, DELIVERING, DELIVERED, COMPLETED
+}
