+++ src/main/java/kitchenpos/orders/domain/OrderType.java
@@ -1,0 +1,5 @@
+package kitchenpos.orders.domain;
+
+public enum OrderType {
+    DELIVERY, TAKEOUT, EAT_IN
+}
