+++ src/main/java/kitchenpos/orders/eatin_orders/domain/OrderTableRepository.java
@@ -1,0 +1,14 @@
+package kitchenpos.orders.eatin_orders.domain;
+
+import java.util.List;
+import java.util.Optional;
+import java.util.UUID;
+
+public interface OrderTableRepository {
+    OrderTable save(OrderTable orderTable);
+
+    Optional<OrderTable> findById(UUID id);
+
+    List<OrderTable> findAll();
+}
+
