+++ src/main/java/kitchenpos/orders/infrastructure/JpaOrderRepository.java
@@ -1,0 +1,10 @@
+package kitchenpos.orders.infrastructure;
+
+import kitchenpos.orders.domain.Order;
+import kitchenpos.orders.domain.OrderRepository;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.UUID;
+
+public interface JpaOrderRepository extends OrderRepository, JpaRepository<Order, UUID> {
+}
